backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "supabase>=2.19.0",
    "uvicorn>=0.36.0",
]
